Introduce timers and lightcontroller
Timers are intended to be a barrier to activate the LED
that students need to adjust via data sent from the Arduino.
They are built around Unity's AnimatorController, whose
state machine control the state of the timers.

From those state changes, the Scripts can update the visuals
in-game and check that all required timers are active.
This is done via the LightController, which will control
the LED and expose a way for the SerialHandler to change
the timer's speeds.

Add the corresponding sprites and animations, which are
also used for the static states of the timer.
Introduce a basic Square prefab for creating basic sprites,
which are not walls.